# docs/api/agents-openapi.yaml
openapi: 3.0.3
info:
  title: Sunny AI - Multi-Agent Learning System API
  description: |
    Sunny AI's Multi-Agent Cognitive Learning Network provides intelligent, adaptive educational experiences
    for children aged 6-10. This API exposes the capabilities of 8+ specialized AI agents working in concert
    to assess, teach, adapt, and support young learners.

    ## Agent Types
    - **Assessment Agent**: Analyzes student performance and identifies knowledge gaps
    - **Path Planner Agent**: Creates adaptive learning paths with prerequisite chains
    - **Content Generation Agent**: Produces personalized lessons, quizzes, and activities
    - **Intervention Agent**: Detects and responds to frustration, confusion, and disengagement
    - **Communication Agent**: Adapts tone, language, and style for emotional intelligence
    - **Reflection Agent**: Generates session summaries and metacognitive prompts
    - **Collaboration Agent**: Facilitates peer learning and social experiences
    - **Game Agent**: Recommends educational games based on engagement and mastery

    ## Architecture Principles
    - **Perception**: Agents perceive student state through interactions and assessments
    - **Cognition**: Orchestrator coordinates agents to make intelligent decisions
    - **Action**: Agents generate recommendations (content, interventions, strategies)
    - **Memory**: Learning states persist across sessions with Bayesian Knowledge Tracing

    ## Performance Targets
...

...
servers:
...
paths:
...
  /agents/reflect:
    post:
      tags: [Reflection]
      summary: Generate reflection/summary
      description: |
        Generates metacognitive reflections:
        - **session_summary**: Post-session insights and achievements
        - **progress_narrative**: Story-format learning journey (week/month/all-time)
        - **self_evaluation**: Prompts for student self-assessment

        Uses Reflection Agent with growth mindset framing.

      operationId: generateReflection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReflectionRequest'
            examples:
              session_summary:
                summary: Post-session summary
                value:
                  student_id: "student-123"
                  session_id: "session-456"
                  reflection_type: "session_summary"

      responses:
        '200':
          description: Reflection generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReflectionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Session not found
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  schemas:
...
    ReflectionRequest:
      type: object
      required: [student_id, session_id, reflection_type]
      properties:
        student_id:
          type: string
        session_id:
          type: string
        reflection_type:
          type: string
          enum: [session_summary, progress_narrative, self_evaluation]
        timeframe:
          type: string
          enum: [session, week, month]
          default: session
    ReflectionResponse:
      type: object
      properties:
        success:
          type: boolean
        reflection:
          type: object
          properties:
            student_id:
              type: string
            session_id:
              type: string
            type:
              type: string
            timestamp:
              type: integer
            content:
              type: object
              description: Reflection content (varies by type)
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Missing required fields"
    RateLimitExceeded:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Rate limit exceeded"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: string
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
security:
  - ApiKeyAuth: []